Cleaned up code, removed unnecessary lines fixed small layout bug
Moved the GameRulesFragment to the main folder as it is not a Dialog
fragment it is a normal fragment.

There was a problem with the words wrapping in funny ways on  my
devices so I changed the width parameters for the text boxes to
wrap_content

Lowered text size to work better on smaller devices.

locked orientation to portrait.